+++ @xen-orchestra/web/src/composables/xo-vm-utils.composable.ts
@@ -1,35 +1,98 @@
-import { useXoVbdCollection } from '@/remote-resources/use-xo-vbd-collection'
-import { useXoVdiCollection } from '@/remote-resources/use-xo-vdi-collection'
-import type { XoVbd } from '@/types/xo/vbd.type'
-import type { XoVm } from '@/types/xo/vm.type'
-import { formatSizeRaw } from '@core/utils/size.util'
-import type { Info, Scale } from 'human-format'
+import { useXoHostCollection } from '@/remote-resources/use-xo-host-collection.ts'
+import { useXoVmCollection } from '@/remote-resources/use-xo-vm-collection.ts'
+import { HOST_POWER_STATE } from '@/types/xo/host.type.ts'
+import { VM_POWER_STATE, type XoVm } from '@/types/xo/vm.type.ts'
+import type { IconName } from '@core/icons'
+import { useTimeAgo } from '@core/composables/locale-time-ago.composable.ts'
+import { useMapper } from '@core/packages/mapper'
+import { parseDateTime } from '@core/utils/time.util.ts'
+import { toComputed } from '@core/utils/to-computed.util.ts'
+import type { MaybeRefOrGetter } from '@vueuse/shared'
+import { computed } from 'vue'
+import { useI18n } from 'vue-i18n'
 
-export function useXoVmUtils() {
-  const { getVbdById } = useXoVbdCollection()
-  const { getVdiById } = useXoVdiCollection()
+export function useXoVmUtils(rawVm: MaybeRefOrGetter<XoVm>) {
+  const { t, locale } = useI18n()
 
-  function getRam(vm: XoVm): Info<Scale<'B' | 'KiB' | 'MiB' | 'GiB' | 'TiB'>> | undefined {
-    return formatSizeRaw(vm.memory.size, 1)
-  }
+  const vm = toComputed(rawVm)
 
-  function getIpAddresses(vm: XoVm) {
-    const addresses = vm.addresses
+  const { getVmHost } = useXoVmCollection()
+  const { isMasterHost } = useXoHostCollection()
 
-    return addresses ? [...Object.values(addresses).sort()] : []
-  }
+  const powerState = useMapper<VM_POWER_STATE, { icon: IconName; text: string }>(
+    () => vm.value.power_state,
+    {
+      [VM_POWER_STATE.RUNNING]: { icon: 'legacy:running', text: t('vm-status.running') },
+      [VM_POWER_STATE.HALTED]: { icon: 'legacy:halted', text: t('vm-status.halted') },
+      [VM_POWER_STATE.PAUSED]: { icon: 'legacy:paused', text: t('vm-status.paused') },
+      [VM_POWER_STATE.SUSPENDED]: { icon: 'legacy:suspended', text: t('vm-status.suspended') },
+    },
+    VM_POWER_STATE.RUNNING
+  )
 
-  function getDiskSpace(vm: XoVm): Info<Scale<'B' | 'KiB' | 'MiB' | 'GiB' | 'TiB'>> | undefined {
-    const vdis = [...vm.$VBDs].map(vbdId => getVbdById(vbdId as XoVbd['id'])?.VDI)
+  const host = computed(() => getVmHost(vm.value))
 
-    const totalSize = vdis.map(vdiId => getVdiById(vdiId)?.size || 0).reduce((sum, size) => sum + size, 0)
+  const isMaster = computed(() => {
+    if (host.value === undefined) {
+      return false
+    }
 
-    return formatSizeRaw(totalSize, 1)
-  }
+    return isMasterHost(host.value.id)
+  })
+
+  const hostPowerState = computed(() => {
+    if (host.value === undefined) {
+      return 'unknown'
+    }
+
+    return host.value.power_state === HOST_POWER_STATE.RUNNING ? 'running' : 'halted'
+  })
+
+  const vmTimeAgo = useTimeAgo(() => new Date(parseDateTime((vm.value.startTime ?? 0) * 1000)))
+
+  const relativeStartTime = computed(() => {
+    if (!vm.value.startTime || vm.value.power_state === VM_POWER_STATE.HALTED) {
+      return t('not-running')
+    }
+    return vmTimeAgo.value
+  })
+
+  const installDateFormatted = computed(() => {
+    if (!vm.value.installTime) {
+      return t('unknown')
+    }
+    return new Intl.DateTimeFormat(locale.value, { dateStyle: 'long' }).format(
+      new Date(parseDateTime(vm.value.installTime * 1000))
+    )
+  })
+
+  const guestToolsDisplay = computed(() => {
+    if (vm.value.power_state !== VM_POWER_STATE.RUNNING) {
+      return { type: 'text', value: '-' }
+    }
+
+    if (!vm.value.managementAgentDetected || !vm.value.pvDriversDetected) {
+      return {
+        type: 'link',
+        value: t('install-guest-tools'),
+        tooltip: !vm.value.managementAgentDetected ? t('management-agent-not-detected') : t('no-tools-detected'),
+      }
+    }
+
+    if (vm.value.pvDriversDetected) {
+      return { type: 'text', value: vm.value.pvDriversVersion || t('installed'), tooltip: t('installed') }
+    }
+
+    return { type: 'text', value: '-' }
+  })
 
   return {
-    getRam,
-    getIpAddresses,
-    getDiskSpace,
+    powerState,
+    host,
+    isMaster,
+    hostPowerState,
+    installDateFormatted,
+    relativeStartTime,
+    guestToolsDisplay,
   }
 }
